--- app/src/main/kotlin/com/blox/payments/ui/registration/birthdate/RegistrationBirthDateUiState.kt
@@ -1,6 +1,0 @@
-package com.blox.payments.ui.registration.birthdate
-
-data class RegistrationBirthDateUiState(
-    val birthDateCompleted: Boolean = false,
-    val birthDateError: Boolean = false
-)
